parser/requirements/base.txt
Scrapy==2.3.0
python-dotenv==0.14.0
psycopg2-binary==2.8.6
scrapy-rotating-proxies==0.6.2
